+++ lib/src/bindings/soloud_controller_ffi.dart
@@ -3,7 +3,6 @@
 import 'dart:ffi' as ffi;
 import 'dart:io';
 
-import 'package:flutter_soloud/src/bindings/bindings_capture_ffi.dart';
 import 'package:flutter_soloud/src/bindings/bindings_player_ffi.dart';
 
 /// Controller that expose method channel and FFI
@@ -20,7 +19,6 @@ class SoLoudController {
                 ? ffi.DynamicLibrary.open('flutter_soloud_plugin.dll')
                 : ffi.DynamicLibrary.process()));
     soLoudFFI = FlutterSoLoudFfi.fromLookup(nativeLib.lookup);
-    captureFFI = FlutterCaptureFfi.fromLookup(nativeLib.lookup);
   }
 
   static SoLoudController? _instance;
@@ -28,6 +26,4 @@ class SoLoudController {
   late ffi.DynamicLibrary nativeLib;
 
   late final FlutterSoLoudFfi soLoudFFI;
-
-  late final FlutterCaptureFfi captureFFI;
 }