+++ lib/src/bindings/soloud_controller_web.dart
@@ -1,7 +1,5 @@
 // ignore_for_file: public_member_api_docs
 
-import 'package:flutter_soloud/src/bindings/bindings_capture_web.dart';
-
 import 'package:flutter_soloud/src/bindings/bindings_player_web.dart';
 
 /// Controller that expose method channel and FFI
@@ -15,8 +13,4 @@ class SoLoudController {
   final FlutterSoLoudWeb _soLoudFFI = FlutterSoLoudWeb();
 
   FlutterSoLoudWeb get soLoudFFI => _soLoudFFI;
-
-  final FlutterCaptureWeb _captureFFI = FlutterCaptureWeb();
-
-  FlutterCaptureWeb get captureFFI => _captureFFI;
 }