+++ test_fixes/soloud_capture_initialization.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_soloud/flutter_soloud.dart';
-
-Future<void> main() async {
-  SoLoudCapture().initCapture(deviceID: 0);
-}